+++ boris-client/src/Boris/Client/Discover.hs
@@ -7,6 +7,7 @@ module Boris.Client.Discover (
 import           Boris.Client.Http (BorisHttpClientError (..))
 import qualified Boris.Client.Http as H
 import           Boris.Core.Data
+import           Boris.Core.Data.Instance
 import           Boris.Representation.ApiV1
 
 import           P
