+++ boris-core/test/Test/Boris/Core/Gen.hs
@@ -4,6 +4,8 @@
 module Test.Boris.Core.Gen where
 
 import           Boris.Core.Data
+import           Boris.Core.Data.Configuration
+import           Boris.Core.Data.Log
 
 import           Data.Text (Text)
 import qualified Data.Text as Text
